Disable overlap check for non-facility rooms
PreventRoomOverlap is a function that tries to resolve a room's overlap issues by potentially swapping it with another room of the same shape. The disableoverlapcheck key makes a room exempt from PreventRoomOverlap's effects. Although the intro room, the pocket dimension, and 1499's dimension are big enough to not actually be swapped with anything, there is no reason they should even be considered for swapping by PreventRoomOverlap, so this commit gives them each the disableoverlapcheck key.