Change version number to 1.3.12
VersionNumber had to be changed to 1.3.12 sooner or later, and with the new save compatibility system, I think this is a good time.

VersionNumber is what appears in the app title and in the corner of the main menu. With the new save compatibility system, it's also the number shown on save files in the load game screen.